worker: introduce ctx to NewResetter
This was directly creating a context.Background(), but we have context's
"from above" that should be used and respected. Generally these contexts
are the context from "func main" or the ctx passed into the Routine
call. For some reason these where nearly always ignored, which I believe
is just due to how people do copy paste when creating a new worker.

I read a bit of code to make sure this was correct (but will have to
rely on real life testing as well). The most interesting case was how we
start up routines concurrently. Before we ignore ctx most of the time =>
if a job failed we didn't shut down quickly. Now we do.

Minor: I also made the Stop implementation respect the passed in
context. From reading usages of ctx and finished this change is correct.

Test Plan: lots of reading of code, otherwise will rely on our tests and
how this behaves on S2. I will only land this change after the release
tomorrow.