chore: Replace QueuedCount -> CountByState with bitset parameter
Previously, the QueuedCount method was confusing because:
1. By default, it actually returned the count for both the 'queued'
   and 'errored' states (despite the name just have 'Queued').
2. There was an additional boolean flag for also returning entries
   in the 'processing' state, but reduced clarity at call-sites.

So I've changed the method to take a bitset instead, mirroring the just-added
Exists API, and renamed the method to a more generic 'CountByState'.

While this does make call-sites a bit more verbose, I think the
clarity win makes the change an overall positive one.