Update tests using mocks for call to Exists
This commit shows the brittle-ness of mock-based testing.
Adding a new method and changing the implementation
should be triggering a compile-time error for a 'fake'
or similar, so that the invariants between various methods
are respected. Checking the history/number of calls does
not actually give any kind of confidence in the
implementation.