refactor: implements a Stream to filter and limit search results
As part of pagination we record the file paths we've so far, so we can
filter them out on the next page. That means we can no longer give
Searcher/Zoekt a static count limit, as we need to stream past already
seen files until we've reached the desired count. That's what this new
Stream implements