database: optional pprof for tenantless_queries
This adds a new profile output for when a sql query does not have a
tenant set. This can be downloaded like other pprof profiles and
inspected by "go tool pprof". This allows us to easily see the most
common stack traces and sources of missing context.

Test Plan: downloaded from the debugserver the new profile type and
inspected with "go tool pprof"