feat(svelte): Support 'got to defintion' for multiple definitions and  'find implementations' (#64349)
This commit adds the necessary logic to open the explorer tab for
multiple references and for implementations. I basically copied what
@camdencheek did for the references tab.


## Test plan

Hover over `adapt` in
http://localhost:5173/github.com/sveltejs/kit@65931f276ac2102032e3032c864a472eee19b7bb/-/blob/packages/kit/src/exports/vite/index.js?L890
and click 'got to definition'. The explore panel with should open with
the definitions tab selected.

I wasn't able to test the find implementations logic because I don't
know for which language we have this implemented.