build(svelte): Build Svelte app with enterprise runset
This commit adds a new command, `web-sveltekit-prod`, which simplies
builds the SvelteKit app.
This command is now run by default when using `sg start`, so the assets
will be available locally.

Note however that the build process is quite slow.

To make the integration a bit more ergonomic, I made additional changes:

- When run via `sg` (determined by checking `DEPLOY_TYPE`) compiler
  warnings are not printed. That reduces noise in the `sg` output.
- I made similar changes to the svelte check bazel target to make it
  easier to spot actuall errors (warnings are still important to fix but
  since they don't cause the target to fail they just make it difficult
  to spot the actual error).
- And since I finally found out about the `onwarn` handler, I made it so
  that warnings about missing declarations are ignored (see the comment
  for why).