fix(svelte): Always close navigation sidebar when navigating to a different page
Problem: Currently the sidebar can reopen in the following scenario:
- Go to search home and submit a search
- On the search results page click the menu icon to open the sidebar and
  click 'Code Search' to go back to search home.
- Execute a different query
- -> this brings you back to search results but with the menu sidebar open

Looking at the code it's understandable that this happens: We never
close the sidebar. But this problem wasn't very visible because every
entry in the nav sidebar either brings you to the React app or to a page
with inline nav menu.

So this commit fixes the issue by closing the sidebar whenever we
navigate to a different page.