tenant: Introduce tenant_id to every table (#64323)
We introduce the tenant_id column to every table without any magic
enforcement or magic default value. Instead we default to NULL and plan
to have an out of band migration. From out testing this change is
instant since it is only a table metadata change. According to the
documentation this sort of ALTER TABLE is instant since Postgres 11.

Test Plan: ran this migration against a clone of the s2 database.

---------

Co-authored-by: Keegan Carruthers-Smith <keegan.csmith@gmail.com>