migration: fix stitch test cases (#64409)
These all fail and are only run outside of bazel (which is why we didn't
notice it). I believe the root cause is google getting more strict with
the URLs you pass it.

Should we consider removing these tests given no one has reported it not
working? Or should we get it working in Bazel?

Test Plan go test ./internal/database/migration/stitch/