migrationtests: Actually run in a transaction when in test mode
When we added a migration that actually relies on the promised property of migrations always running inside a transaction, we noticed that the memoryStore implementation we use in the dbtest package does not actually start a transaction. This implements transactions for the memorystore as well, to make it pass.

Test plan: Our migration now passes, and dbtest performance with and without this change is the same.